// src/app/models/pbx-dialplan-detail.model.ts
import { PbxDomainModel } from './pbx-domain.model';

export class PbxDialplanDetailModel {

  domain_uuid?: string;
  dialplan_uuid?: string;
  dialplan_detail_uuid?: string;
  dialplan_detail_tag?: string;
  dialplan_detail_type?: string;
  dialplan_detail_data?: string;
  dialplan_detail_break?: string;
  dialplan_detail_inline?: string;
  dialplan_detail_group?: string;
  dialplan_detail_order?: string;

  constructor() { }

}